Finished Skinned and Skeletal Animation and inserted basic manipulation functionality.

+++ include/Hydrox/Graphics/Mesh.h
@@ -40,7 +40,7 @@ namespace he
       std::vector<Vector<float, 2>> textureCoords, 
       std::vector<Vector<float, 3>> normals, 
       std::vector<Vector<float, 3>> binormals, 
-      std::vector<Vector<unsigned int, 4>> boneIndices, 
+      std::vector<Vector<float, 4>> boneIndices, 
       std::vector<Vector<float, 4>> boneWeights,
       std::vector<indexType> indices,
       GLuint primitiveType = GL_TRIANGLES);
